backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
python-dotenv
langchain
langchain-groq
langchain-core